refactor: MockerParser 对象如果不传递 buildPath，则不产生 db.json 和本地目录
