+++ lib/mocker/MockerParser.js
@@ -30,16 +30,19 @@ var MockerParser = function () {
     this.basePath = opts.basePath;
     this.matmanMockers = Array.isArray(opts.matmanMockers) ? [].concat(_toConsumableArray(opts.matmanMockers)) : [];
 
-    // 设置默认值和绝对路径
-    this.buildPath = opts.buildPath || './build';
-    if (!path.isAbsolute(this.buildPath)) {
-      this.buildPath = path.resolve(this.basePath, this.buildPath);
+    // 只有传递了 opts.buildPath，才处理 db
+    if (opts.buildPath) {
+      // 设置默认值和绝对路径
+      this.buildPath = opts.buildPath || './build';
+      if (!path.isAbsolute(this.buildPath)) {
+        this.buildPath = path.resolve(this.basePath, this.buildPath);
+      }
+
+      // 注意此处一定要保证存储数据的地址是可存在的，否则会保存失败。
+      fse.ensureDirSync(this.buildPath);
+
+      this.db = store.getDB(path.join(this.buildPath, 'db.json'));
     }
-
-    // 注意此处一定要保证存储数据的地址是可存在的，否则会保存失败。
-    fse.ensureDirSync(this.buildPath);
-
-    this.db = store.getDB(path.join(this.buildPath, 'db.json'));
   }
 
   /**
