+++ test/scripts/mocker/MockerParser.test.js
@@ -2,198 +2,197 @@ const path = require('path');
 const chai = require('chai');
 const expect = chai.expect;
 
-const MockerParser = require('../../../src/mocker/MockerParser');
-const Mocker = require('../../../src/mocker/Mocker');
-const MockModule = require('../../../src/mocker/MockModule');
+const { MockerParser, Mocker } = require('../../../lib');
+const MockModule = require('../../../lib/mocker/MockModule');
 
 describe('./mocker/MockerParser.js', () => {
-    let mockerParser;
+  let mockerParser;
 
-    before(() => {
-        mockerParser = new MockerParser({
-            basePath: path.resolve(__dirname, '../../data/fixtures/mock_server/mockers')
-        });
-
-        // console.log(mockerParser);
+  before(() => {
+    mockerParser = new MockerParser({
+      basePath: path.resolve(__dirname, '../../data/fixtures/mock_server/mockers')
     });
 
-    describe('check basic info', () => {
-        it('should be instanceof MockerParser ', () => {
-            expect(mockerParser).to.be.an.instanceof(MockerParser);
-        });
+    // console.log(mockerParser);
+  });
 
-        it('should contain some fields', () => {
-            expect(mockerParser).to.have.all.keys('basePath', 'matmanMockers');
-        });
+  describe('check basic info', () => {
+    it('should be instanceof MockerParser ', () => {
+      expect(mockerParser).to.be.an.instanceof(MockerParser);
     });
 
-    describe('check demo_01', () => {
-        it('mockerParser.getMockerByName(\'demo_01\') should return correct mocker', () => {
-            let mocker = mockerParser.getMockerByName('demo_01');
+    it('should contain some fields', () => {
+      expect(mockerParser).to.have.all.keys('basePath', 'matmanMockers');
+    });
+  });
 
-            expect(mocker).to.be.an.instanceof(Mocker);
-            expect(mocker.name).to.equal('demo_01');
-            expect(mocker.mockModuleList).to.have.lengthOf(2);
-        });
+  describe('check demo_01', () => {
+    it('mockerParser.getMockerByName(\'demo_01\') should return correct mocker', () => {
+      let mocker = mockerParser.getMockerByName('demo_01');
 
-        it('mockerParser.getMockModuleByName(\'demo_01\', \'error\') should return correct mockModule', () => {
-            let mockModule = mockerParser.getMockModuleByName('demo_01', 'error');
-
-            expect(mockModule).to.be.an.instanceof(MockModule);
-            expect(mockModule.name).to.equal('error');
-
-            return mockModule.getResult()
-                .then((data) => {
-                    expect(data).to.eql({
-                        errCode: 100000
-                    });
-                });
-        });
+      expect(mocker).to.be.an.instanceof(Mocker);
+      expect(mocker.name).to.equal('demo_01');
+      expect(mocker.mockModuleList).to.have.lengthOf(2);
     });
 
-    describe('check demo_02', () => {
-        it('mockerParser.getMockerByName(\'demo_02\') should return correct mocker', () => {
-            let mocker = mockerParser.getMockerByName('demo_02');
+    it('mockerParser.getMockModuleByName(\'demo_01\', \'error\') should return correct mockModule', () => {
+      let mockModule = mockerParser.getMockModuleByName('demo_01', 'error');
 
-            expect(mocker).to.be.undefined;
+      expect(mockModule).to.be.an.instanceof(MockModule);
+      expect(mockModule.name).to.equal('error');
+
+      return mockModule.getResult()
+        .then((data) => {
+          expect(data).to.eql({
+            errCode: 100000
+          });
         });
+    });
+  });
 
-        it('mockerParser.getMockerByName(\'demo_02_renamed\') should return correct mocker', () => {
-            let mocker = mockerParser.getMockerByName('demo_02_renamed');
+  describe('check demo_02', () => {
+    it('mockerParser.getMockerByName(\'demo_02\') should return correct mocker', () => {
+      let mocker = mockerParser.getMockerByName('demo_02');
 
-            expect(mocker).to.be.an.instanceof(Mocker);
-            expect(mocker.name).to.equal('demo_02_renamed');
-            expect(mocker.mockModuleList).to.have.lengthOf(5);
-        });
+      expect(mocker).to.be.undefined;
+    });
 
-        it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'error\') should return correct mockModule', () => {
-            let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'error');
+    it('mockerParser.getMockerByName(\'demo_02_renamed\') should return correct mocker', () => {
+      let mocker = mockerParser.getMockerByName('demo_02_renamed');
 
-            expect(mockModule).to.be.an.instanceof(MockModule);
-            expect(mockModule.name).to.equal('error');
+      expect(mocker).to.be.an.instanceof(Mocker);
+      expect(mocker.name).to.equal('demo_02_renamed');
+      expect(mocker.mockModuleList).to.have.lengthOf(5);
+    });
 
-            return mockModule.getResult()
-                .then((data) => {
-                    expect(data).to.eql({
-                        errCode: 100000
-                    });
-                });
-        });
+    it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'error\') should return correct mockModule', () => {
+      let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'error');
 
-        it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_1\') should return correct mockModule', () => {
-            let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_1');
+      expect(mockModule).to.be.an.instanceof(MockModule);
+      expect(mockModule.name).to.equal('error');
 
-            expect(mockModule).to.be.an.instanceof(MockModule);
-            expect(mockModule.name).to.equal('success_1');
-
-            return mockModule.getResult()
-                .then((data) => {
-                    expect(data).to.eql({
-                        'result': {
-                            'other': 'other',
-                            'result': 1
-                        },
-                        'retcode': 0
-                    });
-                });
-        });
-
-        it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_2\') should return correct mockModule', () => {
-            let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_2');
-
-            expect(mockModule).to.be.an.instanceof(MockModule);
-            expect(mockModule.name).to.equal('success_2');
-
-            return mockModule.getResult()
-                .then((data) => {
-                    expect(data).to.eql({
-                        'result': 2
-                    });
-                });
-        });
-
-        it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_3_renamed\') should return correct mockModule', () => {
-            let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_3_renamed');
-
-            expect(mockModule).to.be.an.instanceof(MockModule);
-            expect(mockModule.name).to.equal('success_3_renamed');
-
-            return mockModule.getResult()
-                .then((data) => {
-                    expect(data).to.eql({
-                        'result': 3
-                    });
-                });
-        });
-
-        it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_4\') should return correct mockModule', () => {
-            let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_4');
-
-            expect(mockModule).to.be.an.instanceof(MockModule);
-            expect(mockModule.name).to.equal('success_4');
-
-            return mockModule.getResult()
-                .then((data) => {
-                    expect(data).to.equal(4);
-                });
-        });
-        it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_4\') with param should return correct mockModule', () => {
-            let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_4');
-
-            expect(mockModule).to.be.an.instanceof(MockModule);
-            expect(mockModule.name).to.equal('success_4');
-
-            return mockModule.getResult({ a: 110 })
-                .then((data) => {
-                    expect(data).to.equal('from_param_110');
-                });
+      return mockModule.getResult()
+        .then((data) => {
+          expect(data).to.eql({
+            errCode: 100000
+          });
         });
     });
 
-    describe('check getMockerByRoute', () => {
-        it('should return demo_01', () => {
-            let mocker = mockerParser.getMockerByRoute('/cgi-bin/a/b/demo_01');
+    it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_1\') should return correct mockModule', () => {
+      let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_1');
 
-            expect(mocker).to.be.an.instanceof(Mocker);
-            expect(mocker.name).to.equal('demo_01');
-        });
+      expect(mockModule).to.be.an.instanceof(MockModule);
+      expect(mockModule.name).to.equal('success_1');
 
-        it('should return null', () => {
-            let mocker = mockerParser.getMockerByRoute('/cgi-bin/a/b/demo_not_exist');
-
-            expect(mocker).to.be.null;
-        });
-
-        it('should return demo_02', () => {
-            let mocker = mockerParser.getMockerByRoute('/cgi-bin/a/b/demo_02');
-
-            expect(mocker).to.be.an.instanceof(Mocker);
-            expect(mocker.name).to.equal('demo_02_renamed');
+      return mockModule.getResult()
+        .then((data) => {
+          expect(data).to.eql({
+            'result': {
+              'other': 'other',
+              'result': 1
+            },
+            'retcode': 0
+          });
         });
     });
 
-    describe('check getResInfoByRoute', () => {
-        it('demo_01 ok', () => {
-            let resInfo = mockerParser.getResInfoByRoute('/cgi-bin/a/b/demo_01');
+    it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_2\') should return correct mockModule', () => {
+      let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_2');
 
-            expect(resInfo).to.have.all.keys('mockerItem', 'mockModuleItem', 'moduleFullPath', 'params');
-            expect(resInfo.mockerItem.name).to.equal('demo_01');
-            expect(resInfo.mockModuleItem.name).to.equal('error');
-        });
+      expect(mockModule).to.be.an.instanceof(MockModule);
+      expect(mockModule.name).to.equal('success_2');
 
-        it('should return null', () => {
-            let mocker = mockerParser.getResInfoByRoute('/cgi-bin/a/b/demo_not_exist');
-
-            expect(mocker).to.be.null;
-        });
-
-        it('demo_02 ok', () => {
-            let resInfo = mockerParser.getResInfoByRoute('/cgi-bin/a/b/demo_02');
-
-            expect(resInfo).to.have.all.keys('mockerItem', 'mockModuleItem', 'moduleFullPath', 'params');
-            expect(resInfo.mockerItem.name).to.equal('demo_02_renamed');
-            expect(resInfo.mockModuleItem.name).to.equal('success_1');
+      return mockModule.getResult()
+        .then((data) => {
+          expect(data).to.eql({
+            'result': 2
+          });
         });
     });
+
+    it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_3_renamed\') should return correct mockModule', () => {
+      let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_3_renamed');
+
+      expect(mockModule).to.be.an.instanceof(MockModule);
+      expect(mockModule.name).to.equal('success_3_renamed');
+
+      return mockModule.getResult()
+        .then((data) => {
+          expect(data).to.eql({
+            'result': 3
+          });
+        });
+    });
+
+    it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_4\') should return correct mockModule', () => {
+      let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_4');
+
+      expect(mockModule).to.be.an.instanceof(MockModule);
+      expect(mockModule.name).to.equal('success_4');
+
+      return mockModule.getResult()
+        .then((data) => {
+          expect(data).to.equal(4);
+        });
+    });
+    it('mockerParser.getMockModuleByName(\'demo_02_renamed\', \'success_4\') with param should return correct mockModule', () => {
+      let mockModule = mockerParser.getMockModuleByName('demo_02_renamed', 'success_4');
+
+      expect(mockModule).to.be.an.instanceof(MockModule);
+      expect(mockModule.name).to.equal('success_4');
+
+      return mockModule.getResult({ a: 110 })
+        .then((data) => {
+          expect(data).to.equal('from_param_110');
+        });
+    });
+  });
+
+  describe('check getMockerByRoute', () => {
+    it('should return demo_01', () => {
+      let mocker = mockerParser.getMockerByRoute('/cgi-bin/a/b/demo_01');
+
+      expect(mocker).to.be.an.instanceof(Mocker);
+      expect(mocker.name).to.equal('demo_01');
+    });
+
+    it('should return null', () => {
+      let mocker = mockerParser.getMockerByRoute('/cgi-bin/a/b/demo_not_exist');
+
+      expect(mocker).to.be.null;
+    });
+
+    it('should return demo_02', () => {
+      let mocker = mockerParser.getMockerByRoute('/cgi-bin/a/b/demo_02');
+
+      expect(mocker).to.be.an.instanceof(Mocker);
+      expect(mocker.name).to.equal('demo_02_renamed');
+    });
+  });
+
+  describe('check getResInfoByRoute', () => {
+    it('demo_01 ok', () => {
+      let resInfo = mockerParser.getResInfoByRoute('/cgi-bin/a/b/demo_01');
+
+      expect(resInfo).to.have.all.keys('mockerItem', 'mockModuleItem', 'moduleFullPath', 'params');
+      expect(resInfo.mockerItem.name).to.equal('demo_01');
+      expect(resInfo.mockModuleItem.name).to.equal('error');
+    });
+
+    it('should return null', () => {
+      let mocker = mockerParser.getResInfoByRoute('/cgi-bin/a/b/demo_not_exist');
+
+      expect(mocker).to.be.null;
+    });
+
+    it('demo_02 ok', () => {
+      let resInfo = mockerParser.getResInfoByRoute('/cgi-bin/a/b/demo_02');
+
+      expect(resInfo).to.have.all.keys('mockerItem', 'mockModuleItem', 'moduleFullPath', 'params');
+      expect(resInfo.mockerItem.name).to.equal('demo_02_renamed');
+      expect(resInfo.mockModuleItem.name).to.equal('success_1');
+    });
+  });
 });
